COMMONS-552 : make sure PortalContainer has been created before running tests
The test org.exoplatform.commons.persistence.impl.ExoTransactionalAnnotationTest uses the EntityManagerService which is declared in portal container configuration.
The test class does not explicitly creates the PortalContainer. So if another test class executed before this test class creates the Portal Container (case of local build), the Portal Container is already created and the EntityManagerService is available, and the test passes.
But if no test class creating the PortalContainer is run before this one, the PortalContainer is not created and the EntityManagerService is not instantiated, so the test fails.

The fix makes sure the PortalContainer is created for this test class.