SOC-6069 : always set new and updated notifications as unread and in popover (#232)
When a notification is updated (for example when several comments are done on the same activity), the template builder (for example the one from social : org.exoplatform.social.notification.channel.template.WebTemplateProvider) can updates the attributes of the notification.
The problem is when it updates the unread and onPopover attributes of the notification. If it changes them, the updated notification can be stored as read and/or not onPopover, whereas we always want an updated notification to be shown in the popover as unread.
This fix set explicity these 2 attributes to make sure the updated notification is unread and in popover, no matter what the template builder will update.